[types] Move module bindings from rooch-framework to rooch-types
