+++ crates/rooch-framework-tests/src/tests/ed25519_validator_tests.rs
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use moveos_types::transaction::MoveAction;
-use rooch_framework::bindings::empty::Empty;
 use rooch_key::keystore::{AccountKeystore, InMemKeystore};
+use rooch_types::framework::empty::Empty;
 use rooch_types::{
     crypto::BuiltinScheme,
     transaction::{rooch::RoochTransactionData, AbstractTransaction},
@@ -15,8 +15,7 @@ use crate::binding_test;
 fn test_validate() {
     let binding_test = binding_test::RustBindingTest::new().unwrap();
     let ed25519_validator = binding_test
-        .as_module_bundle::<rooch_framework::bindings::ed25519_validator::Ed25519Validator>(
-    );
+        .as_module_bundle::<rooch_types::framework::ed25519_validator::Ed25519Validator>();
 
     let keystore = InMemKeystore::new_ed25519_insecure_for_tests(1);
     let sender = keystore.addresses()[0];
