--- crates/rooch-framework/src/bindings/mod.rs
@@ -1,12 +1,0 @@
-// Copyright (c) RoochNetwork
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod address_mapping;
-pub mod auth_validator;
-pub mod ecdsa_k1_recoverable_validator;
-pub mod ecdsa_k1_validator;
-pub mod ed25519_validator;
-pub mod empty;
-pub mod schnorr_validator;
-pub mod session_key;
-pub mod transaction_validator;
