+++ crates/rooch-framework/src/lib.rs
@@ -1,7 +1,6 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod bindings;
 pub mod natives;
 
 pub use rooch_types::addresses::*;
