+++ crates/rooch-types/src/framework/address_mapping.rs
@@ -1,7 +1,8 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::ROOCH_FRAMEWORK_ADDRESS;
+use crate::address::{MultiChainAddress, RoochAddress};
+use crate::addresses::ROOCH_FRAMEWORK_ADDRESS;
 use anyhow::Result;
 use move_core_types::{account_address::AccountAddress, ident_str, identifier::IdentStr};
 use moveos_types::{
@@ -11,7 +12,6 @@ use moveos_types::{
     transaction::FunctionCall,
     tx_context::TxContext,
 };
-use rooch_types::address::{MultiChainAddress, RoochAddress};
 
 /// Rust bindings for RoochFramework address_mapping module
 pub struct AddressMapping<'a> {
