+++ crates/rooch-types/src/framework/mod.rs
@@ -1,7 +1,15 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-/// types mapping from Framework Move types to Rust types
-//TODO should be in the framework crate?
+/// Types mapping from Framework Move types to Rust types
+/// Module binding for the Framework
+//TODO there modules maybe generated by ABI in the future
+pub mod address_mapping;
 pub mod auth_validator;
+pub mod ecdsa_k1_recoverable_validator;
+pub mod ecdsa_k1_validator;
+pub mod ed25519_validator;
+pub mod empty;
+pub mod schnorr_validator;
 pub mod session_key;
+pub mod transaction_validator;
