+++ crates/rooch/src/commands/account/commands/update.rs
@@ -10,13 +10,13 @@ use move_core_types::{
 };
 use moveos_types::module_binding::ModuleBundle;
 use moveos_types::{move_types::FunctionId, transaction::MoveAction};
-use rooch_framework::bindings::{
+use rooch_key::keystore::AccountKeystore;
+use rooch_rpc_api::jsonrpc_types::ExecuteTransactionResponseView;
+use rooch_types::framework::{
     ecdsa_k1_recoverable_validator::EcdsaK1RecoverableValidator,
     ecdsa_k1_validator::EcdsaK1Validator, ed25519_validator::Ed25519Validator,
     schnorr_validator::SchnorrValidator,
 };
-use rooch_key::keystore::AccountKeystore;
-use rooch_rpc_api::jsonrpc_types::ExecuteTransactionResponseView;
 use rooch_types::{
     address::RoochAddress,
     crypto::{BuiltinScheme, PublicKey},
